third_party/blink/web_tests/external/wpt/css/css-backgrounds/parsing/webkit-border-radius-valid-expected.txt

This is a testharness.js-based test.
[FAIL] Serialize -webkit-border-radius from border-radius: 1px 2px
  assert_array_equals: expected property 0 to be "border-top-left-radius: 1px" but got "border-top-left-radius: 1px 2px" (expected array ["border-top-left-radius: 1px", "border-top-right-radius: 2px", "border-bottom-right-radius: 1px", "border-bottom-left-radius: 2px"] got ["border-top-left-radius: 1px 2px", "border-top-right-radius: 1px 2px", "border-bottom-right-radius: 1px 2px", "border-bottom-left-radius: 1px 2px"])
Harness: the test ran to completion.
